Make the <k> tag mode-aware.
Previously there was no way to produce links to the appropriate mode-specific
binding of a key using the <k> tag (e.g., <Tab> in the documentation always
linked to the Normal-mode command, although its command-line functionality is
different).  This commit introduces a `mode' attribute to the <k> tag, making
the distinction possible.