Fixed Array comprehensions in noscript.js
The array comprehension syntax changed in FF 46. Where before you would
do, `[X for (Y in Z)]` now you do `[for (Y of Z) Y]`.

I've also changed the destructuring assignment on line 314, as they
aren't supported, and stop the variables from being bound. See [here]
for more information on that. Now we bind to a single variable, and pull
out the first and second items as needed.

[here] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Array_comprehensions#SpiderMonkey-specific_implementation_notes